Fix dependency cruiser issue to avoid circular imports
