+++ packages/liveblocks-react/src/index.ts
@@ -17,7 +17,7 @@ export { shallow, isNotificationChannelEnabled } from "@liveblocks/client";
 
 // Export all the top-level hooks
 export { RegisterAiKnowledge, RegisterAiTool } from "./ai";
-export type { RegisterAiKnowledgeProps, RegisterAiToolProps } from "./ai";
+export type { RegisterAiKnowledgeProps, RegisterAiToolProps } from "./types/ai";
 export { ClientContext, RoomContext, useClient } from "./contexts";
 export {
   createLiveblocksContext,
